<==== Mercurial data ====>
Branch: feature_305_ldap

+++ mediaserver_core/src/streaming/hls/hls_live_playlist_manager.h
@@ -46,7 +46,7 @@ namespace nx_hls
 
         void clear();
         //!Time (millis) from last usage of this object
-        size_t inactivityPeriod() const;
+        qint64 inactivityPeriod() const;
 
     private:
         MediaStreamCache* const m_mediaStreamCache;
